+++ tests/unittests/test_indicators_definitions.py
@@ -1,4 +1,21 @@
+import asyncio
+from unittest.mock import AsyncMock
+
+import geojson
+import pytest
+
 from ohsome_quality_api.indicators import definitions, models
+
+
+@pytest.fixture(scope="class")
+def mock_select_eubucco_coverage(class_mocker, feature_germany_berlin):
+    async_mock = AsyncMock(
+        return_value=[{"geom": geojson.dumps(feature_germany_berlin)}]
+    )
+    class_mocker.patch(
+        "ohsome_quality_api.indicators.building_completeness.indicator.db_client.get_eubucco_coverage",
+        side_effect=async_mock,
+    )
 
 
 def test_get_indicator_names():
@@ -24,3 +41,12 @@ def test_get_indicator_definitions_with_project():
     for indicator in indicators.values():
         assert isinstance(indicator, models.IndicatorMetadata)
         assert indicator.projects == ["core"]
+
+
+def test_get_coverage(mock_select_eubucco_coverage):
+    coverage = asyncio.run(definitions.get_coverage("building-comparison"))
+    assert coverage.is_valid
+    assert isinstance(coverage, geojson.FeatureCollection)
+    coverage = asyncio.run(definitions.get_coverage("mapping-saturation"))
+    assert coverage.is_valid
+    assert isinstance(coverage, geojson.FeatureCollection)